pool: Delete files concurrently
Motivation:

cleaner submits batches of files to delete to a pool, but a pool processed this
batch sequentially. On a pool with slow meta data operations (for whatever
reason), this can result if fairly low deletion rates and this can in turn slow
down deletion in the entire system.

Modification:

Submit delete to the pool's worker thread pool. Not only does this free one of
the message threads, it allows the delete tasks to be run concurrently. The
increase in concurrency can reduce the overhead of syncing the meta data to
disk after every deletion, thus increasing the deletion rate.

Result:

Improved the deletion rate of files being cleaned on a pool by running deletion
of multiple files on a pool concurrently from the pool's worker thread pool.

I am asking for backport as sites have complained about deletion throughput,
but will only ask for backport to 2.16 as it isn't really fixing a bug. Sites
should move ahead to 2.16 if they want to benefit from this improvement.

Target: master
Request: 3.0
Request: 2.16
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9924/