ftpclient: encrypt SITE CLIENTINFO command
Motivation:

The ftpclient is used by dCache for third-party copies and for the
client software.  This client does not work with a Globus FTP server.

The ftpclient sends a SITE CLIENTINFO command after authentication, but
sends the command unencrypted.  The dCache and Globus FTP servers behave
differently, with the latter disconnecting the client if it receives
such an unencrypted command.

Modification:

The ftpclient sends a SITE CLIENTINFO unencrypted, which triggers the
Globus FTP server to disconnect the client.  With this patch, that
command is sent encrypted.

Result:

The ftpclient works with Globus FTP server.

Target: master
Request: 3.0
Request: 2.16
Request: 2.15
Request: 2.14
Requires-notes: yes
Requires-book: no
Requires-srmclient-notes: yes
Patch: https://rb.dcache.org/r/9934/
Acked-by: Tigran Mkrtchyan