Motivation:
Returning the list of children of a directory is an O(n) operation.
Currently, when a user request for the children of a directory through
the restful-api, all the children will be returned. This can hinder the
performance of an application if the returned data is very large.

Modification:

This patch adds two query parameters namely: limit and offset. The limit
and offset parameter takes in integer values that will be use to set the
requested range. These can only be used in combination with the children
query parameter.

Result:

User can now choose the amount of data returned from listing a directory.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9992/