resilience: fix fairness algorithm bugs in file operation map
Motivation:

In order to provide some fairness in handling requests for file copy
or removal, the resilience file operation map uses an algorithm which
allocates free slots to queued requests on the basis of three factors:

    whether foreground vs background (i.e., new files vs files from a location scan);
    what the total proportion of those two types to the total waiting requests is;
    what the maximum allocation percentage is (this is configurable).

The algorithm attempts to promote requests to running on the basis of
the backlog (more requests for the queue with the greater backlog),
up to a maximum (so that starvation is avoided).

There are several bugs in the current code, however.  Aside from
a (silly) arithmetic error (interpreting max allocation weight on
a scale of 0 < p < 100 instead of 0 < p < 1), a number of different
edge cases were not being handled properly.

Modification:

Add the necessary adjustments and restructure somewhat the main method.
I decided to factor out this code into a separate class, though I
did not expose its selection through Spring injection.  The refactoring
is only in case we should discover the need to allow for customizing
the algorithm in the future.  The code changes should remain
invisible to the user.

The allocation algorithm was in need of a minimal unit test, which
has been provided.

Result:

Correct algorithm behavior.

Target: master
Request: 3.0
Request: 2.16
Acked-by: Tigran