lm: Log when connectors are killed
Motivation:

Symmetry with the creation event.

Modification:

Added a log statement in the location manager indicating when a connector
cell is killed.

Result:

For symmetry with the logging performed when location manager created a
connector, a log message was added for when the connector is killed.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/10012/