Fix class path generation
Motivation:

printClassPath calls printPluginClassPath, but since printPluginClassPath redefines
the classpath variable without marking it as a local variable, the resulting path
is incorrect.

Modification:

Mark classpath as a local variable and explicitly initialize it to prevent it from
inheriting the value of the outer scope.

Also stripped empty paths from the classpath. This in turn required proper suppression
of glob resolution in the shell to avoid that the classpath gets expanded - previously
the superflous inclusion of a leading colon caused the expansion to fail and thus the
literal path was included (ie it worked by accident).

Result:

Fixed a bug that caused duplicate entries in the Java classpath.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10070/