resilience: restore inaccessible file handler interface
Motivation:

An earlier version of Resilience provided a configurable api
in the form of a handler interface for dealing with files
which at the time of verification had no accessible, i.e.,
readable replicas.   This was eliminated in favor of a
simple in-lined call to raise an alarm.

However, it may be that future versions of dCache
require or prefer special handling of such cases.
One such case might, for instance, attempt to restore
a CUSTODIAL ONLINE file from tape onto a writable pool.

Modification:

The interface has been restored, and the default behavior
implemented as a concrete subclass. (Note: there is actually
a minor bug fix incorporated here, as the current handler
could actually try to copy, a subsequently fail, a CUSTODIAL
ONLINE file without any readable source replicas.  A
separate, backportable patch will be provided for this fix
to other branches.)

Result:

Ability to plug in specialized implementations of the
inaccessible file handling behavior.

Target: master
Require-notes: no
Require-book:  no
Acked-by: Dmitry Litvintsev (litvinse@fnal.gov)