chimera-nfs: add get and set for storage uri
Motivation:

The enstore HSM client currently stores relevant information in layer 1
and 4 for each file written to tape.  At the same time, this information
is also stored in the t_locationinfo table in the form of a URI, resulting
in data duplication.  Eliminating layer 1 and 4 data would free up
considerable space in Chimera.

Modification:

A dot command, ".(suri)(<file>)", has been added to NFS by which
one can store, retrieve and modify the URI string(s).  A normal user
can read and write once to this file.  Root can overwrite and append
as well.  Deletion is achieved by opening and truncating the file
(all location entries of type TAPE are deleted from the table whenever
the file size is set to 0).

In addition to the new FsInode, we have added catch clauses to two
methods in ChimeraVfs in order to handle PermissionDenied exceptions
correctly.   We have also added a new method, clearTapeLocations,
to the file system API, in order to encapsulate the specific forced
delete of only type TAPE locations.

A unit test is provided which checks that permissions are enforced.
The URI string written to this file is passed first through a
validity check for both URI syntax and structure; the URI must
be hierarchical and have a scheme, host and query part in order
to pass the check.

Result:

The location URI for type TAPE can now be added directly to
the t_location table; reliance on layer information
and on the ".(use)..." commands is thereby discontinued.

See further the testing notes below.

Target: master
Acked-by: Tigran
Require-book: yes
Require-notes: yes