lm: Do not rely on thread interrupt for shutdown
Motivation:

We have repeatedly tried to fix a problem in which the thread interrupt issued
during connector shutdown appears to get lost. The problem is still not fixed.
This is apprently not a reliable method of shutting down the location manager
connector thread.

Modification:

Switch to a volatile boolean field as a shutdown indicator. A thread interrupt
is still issued to wake up the connector thread, but the shutdown will complete
even if the interrupt is lost (at worst, we have to wait for a connection
timeout or for a connection retry).

Result:

Hopefully finally fixed the issue with duplicate location manager connectors
caused by failure to shutdown a connector cell upon topology changes (e.g. core
comain or zookeeper restart).

Target: trunk
Request: 3.0
Request: 2.16
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10099/