ftp: add support for dynamic checksum calculation
Motivation:

Globus transfer service supports a "verify file integrety after
transfer" option, which is enabled by default.  When enabled, the MD5
checksum values from source and destination files are compared to ensure
the transferred files are not corrupted.

Other clients (typically from grid communities) support similar
integrety verification, but require the storage system provide ADLER32
checksums.

Currently dCache pools calculate, at most, a single checksum on data
ingress.  By default this is ADLER32, but a work-around means that, for
uploads by Globus transfer service, pools will calculate the MD5
checksum.

While this work-around allows Globus uploads to succeed when verifying
file integrety, it does not address the interoperability problem between
clients requiring ADLER32 and those requiring MD5 checksums.

Modification:

Add support in FTP door to calculate an additional checksum value.  This
is achieved by reading the content of the file and calculating two
checksums: the algorithm of an already known checksum and that of the
desired checksum.  If the checksums and file sizes match then the
desired checksum is accepted, stored in Chimera and returned to the
client.

The work-around (where the pool calculates MD5 checksums for Globus
transfer service) is removed.  A client requesting ADLER32 from some
non-FTP door will not trigger any dynamic checksum calculation.  As
Globus transfer service only uses FTP, removing the work-around should
provide the best compatibility.

Note that this process of calculating the checksum is sub-optimal.  It
transfers data to the door rather than calculating it internally within
the pool.  It also does not take advantage of the pool's ability to
calculate checksum on-the-fly.  However, the in-door mechanism has the
advantage of only requiring the door be updated, so back-porting becomes
non-invasive.

It is anticipated that future versions of dCache pool will support
calculating multiple on-transfer checksums and in-place dynamic checksum
calculation.  When these changes in place, the in-door dynamic checksum
calculation should be removed.

Result:

Support Globus transfers of any dCache file with the "verify file
integrety after transfer" option enabled.

Target: master
Request: 3.1
Request: 3.0
Request: 2.16
Requires-notes: yes
Requires-book: no