restful-api: add file mime-type to file attribute
Motivation:

At the moment, only the file-mime-type of the children of a requested
path are returned (provided the path has children). Hence when a
metadata of particular path (to a file or directory) is requested for,
the returned result look like this:
```js
{
  "size": 512,
  "creationTime": 1490344716696,
  "fileType": "DIR",
  "mtime": 1490344985546
}
```

Since, file-mime-type is heavily used by dcache-view, it is desirable
to see the file file-mime-type of the requested path included in the
returned result.

Modification:

Include the setting of file-mime-type inside `chimeraToJsonAttributes`
method and the restful-api path method responsible for returning the
file attribute json was adjusted to reflect this change.

Result:

The new response will look like this:
```js
{
  "fileMimeType": "application/vnd.dcache.folder",
  "size": 512,
  "creationTime": 1490344716696,
  "fileType": "DIR",
  "mtime": 1490344985546
}
```

Target: trunk
Request: 3.1
Require-note: no
Require-book: no
Acked-by: Tigran Mkrtchyan

Reviewed at https://rb.dcache.org/r/10139/