dcache-restful-api: add transfer service implementation
Motivation:

Service should (a) periodically collect and cache transfer information;
(b) service client requests for either a full list of transfers or for
a single transfer; (c) allow clients to 'page' through a set of
transfers according to offset and limit.

Modification:

Service is implemented using the collector to populate a map.  Client
requests are mapped to a cache of snapshots each identified by a UUID.
If the client request lacks the id, a new snapshot is recorded and
a new UUID generated.  Snapshots expire after a designated time following
the last read/access of the snapshot.  The maximum size of the cache is
determined by a property.

Admin commands have been implemented to list the current transfers;
filtering on various fields is possible.  One can also reset timeout
properties or interrupt the wait on the collect cycle to force
an immediate update of the active transfer map.

A few additions to the TransferInfo and UserInfo objects have been
made for the sake of convenience.

Result:

Service is fully implemented with this patch.

Target: master
Depends-on: #10086
Depends-on: #10116
Depends-on: #10117
Acked-by: Tigran