common: move billing TimeFrame class to the common histograms package
Motivation:

In order to support the delivery of histogram data to the frontend,
and then by the frontend to clients via RESTful apis, a set of
histogram abstractions will be created in the common package.

Originally,  histogram classes were used only by billing
and were prepared using the TimeFrame utility class.

For billing plots, this class is still useful, so we wish
to move it into the common package in order to provide
backward support for this billing construct. (The time frame
abstraction, of course, need not be limited to billing data.)

Modification:

Move TimeFrame out of billing and into the new package.

Result:

Some classes have new imports.  These classes will largely
disappear once the old web pages do.

Target: master
Acked-by: Tigran