common: add general-purpose histogram and histogram models, with unit tests
Motivation:

Several frontend dCache admin services are projected to serve,
as part of their RESTful API, histogram data which can be used
by the client to generate plots.

There is thus a need for a few common types for generating
and delivering time series or other forms of data which
can be then visualized by the client as a 2D graph of some
sort.

Modification:

This patch adds several classes to the common module.

Histogram is largely just a Java bean carrying
a 2D array of doubles, along with a few other fields
providing some information necessary for its
interpretation.

HistogramModel and its subtypes (counting, timeseries)
is a service-side datatype which can be used to
organize, store and update the histogram data.  The
counting histogram bins the raw values by counting
how many fall within a given range. The number of bins
is fixed and their width is computed based on the
range of values to be stored.  The time series
histogram provides a moving window of date-time
bins and can be updated; values falling beyond the
highest bin cause the buffer to be rotated.  Updates
can either sum, replace, or be averaged with the
prior value in that bin.  This histogram model
also supports construction from the TimeFrame class.

In order to be able to update using running averages,
update counts on a bin-by-bin basis must be
maintained.  This, along with basic statistical
counters, are maintained by the HistogramMetadata
object composed into the model.

The HistgramModel is converted to the generic Histogram
via the toHistogram method.

A set of unit tests has been provided.

Result:

Histogram abstraction and construction which is reusable.

Target: master
Acked-by: Tigran