system-test: update credential-generating script
Motivation:

The Disposable CA is generated for the system-test credentials and
destroyed after, allowing us to trust it.  Unfortunately, the DN is
currently always the same, which means the subject hash (as used in an
OpenSSL -style trust-store) is constant between updated.

This is undesirable as it becomes more involved discovering whether a
particular Disposable CA instance is trusted.  This makes diagnosing
problems unnecessarily time-consuming.

Modification:

Include the date in the Disposable CA's DN.

Result:

Different hash for each run of the Disposable CA, making it easier to
diagnose any problems.

Target: master
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/10187/
Acked-by: Tigran Mkrtchyan