xrootd: refactor read-write logic for pool open requests
Motivation:
The reason why pool responds to similar requests with different error code is not obvious.

Change:
Rewrite the if tree. Separate the file state and the request flags to see the possible cases more clearly.
Add comments, especially for non-obvious cases.
Maybe some of the non-obvious cases are wrong and should be changed within the same pull request?

Result:
The logic is equivalent so the result should be exactly the same in all cases.