http: fix non-/ root for WebDAV and frontend doors
Motivation:

dCache supports the concept of non-/ root; this is equivalent to the
unix 'chroot' command.  Typically, the non-/ root directory can come
from a (static) configuration of the door, or from users who have some
user-specific non-/ root directory, or both.

When both the door and the user have a root directory then it is
possible to configure dCache so the user is unable to use that door;
e.g., door has root /data/experiment-1 and user has root
/data/experiment-2 then it is impossible for the user to use the door
without relaxing one of the root conditions.

The WebDAV door currently supports static root directory but does not
support users with non-/ root directories.

The frontend door provides some support for a non-/ door root, but this
support is incomplete.  As with the WebDAV door, the frontend provides
no support for users with non-/ root.

Modification:

Update PathMapper class so it can check whether a user's root is
compatible with the door's root.  To be compatible, the two roots must
be equal or one must be a sub-path of the other.  The 'effective root'
is the longest path of the two roots.

The methods responsible for mapping request-paths to dCache-paths and
vice versa (dCache-path to request-path) now use the effective root.

Add support in PathMapper to calculate relative paths that ensure that
the path cannot walk outside the effective root.

Some basic refactoring to simply the patch.

Result:

Users with non-/ root directory will see the same files and directories
under WebDAV as with other protocols.

Frontend provides information consistent with other protocols for users
with a non-/ root.

Frontend 'mkdir' and 'mv' operations are updated to honour door and user
roots.

Target: master
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/10214/
Acked-by: Tigran Mkrtchyan