ftp: add support for SITE USAGE command
Motivation:

There is a semi-standard mechanism to discover capacity usage in FTP:
the SITE USAGE command.  WLCG wishes to use this to discover
space-reservation information.

Modification:

Extract the existing code for caching this information in SRM
as utility methods for building such caches.

Update LoginBroker to inject spacemanager location.

Update the NettyLineBasedDoor to build caches for talking
with SpaceManager that are shared between all FTP sessions
and that are injected into AbstractFTP door.

Add support for SITE USAGE command in AbstractFTP door that
queries SpaceManager (via caches).

If the optional TOKEN argument is supplied then the value is
used to discover reservations associated with the current
user that have this description.  The path is ignored.

If the optional TOKEN argument is omitted then the path is
examined to discover if there is a WriteToken tag.  If so,
this is used.

Result:

The SITE USAGE command is implemented, supplying information
from SpaceManager using reservations with a description that
matches the TOKEN argument (if a TOKEN was given) or that are
bound to the supplied path (if TOKEN was omitted).

Target: master
Patch: https://rb.dcache.org/r/10226/
Request: 3.1
Request: 3.0
Request: 2.16
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan