frontend,webdav: add supress-wwwauthenticate to allow headers
Most (if not all) browsers do a prefligt request and if one of the setted
headers is not found in the  Access-Control-Allow-Headers response header;
the main request response will not be pass down.

Hence, this patch add Suppress-WWW-Authenticate to the list of allowed headers.

This suppose to be part of https://rb.dcache.org/r/10227/ but I forgot to
include it :(.

Target: trunk
Request: 3.0
Request: 3.1
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10233/