frontend: refactor static (configuration) data
Motivation:

Frontend exposes some static information, mainly to supply dCache-View
with configuration information.  It does this by building (during dCache
start-up) a small JavaScript script that declares the CONFIG variable as
a simple JSON object.  The frontend then replies with this content when
a client makes a GET request to some hard-coded path.

This is problematic because a) which information is exposed is not
obvious, b) it is hard for admins to add additional information, c) it
is possible for an admin to configure dCache so that dCache emits
malformed JSON.

Modification:

Use a configuration prefix to identify properties that should be exposed
via JSON object.

Add a Handler to write out this data as either a JSON object or as an
equivalent JavaScript script, based on HTTP content negotiation,
otherwise from the "filename" extension (.js or .json), otherwise
default to JavaScript.

The JSON object keys have been rationalised, so that those key-value
pairs intended for dCache-View are clearly indicated as such.  The
existing documentation has been collated.  The existing keys are also
publish for backwards compatibility.

This version of the patch retains support for existing config options,
but marks them as deprecated.

This version of the patch has been modified when back-porting to match
the older version of Jetty.

Result:

Easily configurable JSON is available via frontend.

Target: master
Request: 3.1
Request: 3.0
Request: 2.16
Require-notes: yes
Require-book: yes
Fixes: #3231
Patch: https://rb.dcache.org/r/10282/
Acked-by: Albert Rossi

Conflicts:
	skel/share/services/frontend.batch

Conflicts:
	skel/share/defaults/frontend.properties
	skel/share/services/frontend.batch