pool: refactor pool to use StandardOpenOption instead of IoMode
Motivation:
We have our own IO mode to indicate read or write modes of a mover.
Since Java7, there is a standard option to indicate that.

Modification:
The IoMode class replaces with set of StandardOpenOption. The set is used
for two reasons:
 - standard FileChannel needs a set of options, like WRITE + CREATE
 - it will allow in the future to indicate truncate or append

Result:
No visible changes for end users. Less code to maintain, step in a
direction of modifiable files.

Acked-by: Paul Millar
Target: master
Require-book: no
Require-notes: no