authz: add initial support for Roles
Motivation:

It may be desirable for dCache to behave differently for the same user
between different login session.  For example, an administrative user
may wish sometimes to interact with dCache as a normal user (for
testing purposes or to reduce the risk of mistakes), and other times to
interact with dCache with effectively root-like prividges.

Modification:

Introduce the concept of roles.  A user may log in requesting zero or
more roles and the login process chooses which roles (if any) to enact.

For the HTTP-based services (webdav and restful), this patch introduces
an extension to allow users to choose which roles they wish to enable.
This is achieved by appending '#' and a comma-separated list of roles to
their username; for example, user "fred" wishing to adopt "role-a" and
"role-b" roles would authenticate with the username
"fred#role-a,role-b".

This patch introduces initial, limited support for the "admin" role.

The "admin" role is intended to give users root-like permissions without
the user knowing some (shared) root credential and with the user's
identity still being bound to that user.  The result is that the user
with "admin" role is able to do everything, but files created by that
user are still owned by that user and log entries show which user
enacted changes.

An "admin" role user has a root directory of '/' and no restrictions.

There are several limitations with the current implementation: as Roles
are LoginAttributes, it is only available in the door.  Therefore doors
must detect when the user has the admin Role and send the root identity.
Therefore when handling requests for an admin-Role user, doors must
choose which operations to send the user's real identity and which to
send ROOT.

The logic for authorizing a user's adoption of a Role lies in a new
gPlazma module: 'roles'.  This new code has no effect without adding
this module within gPlazma configuration.

Result:

A user can request "admin" role when authenticating with the webdav or
restful services.  If authorized to assert that role, the user then has
additional authorisation within dCache.

Target: master
Require-notes: yes
Require-book: yes
Patch: https://rb.dcache.org/r/10281/
Acked-by: Albert Rossi
Acked-by: Tigran Mkrtchyan