common: use round instead of floor to bin counting histogram values
Motivation:

It seems less confusing, when comparing actual values
to the binned values in a counting histogram, to bin
using the rint function rather than floor; in this way,
values on both sides of the bin's lower bound cluster
to that bin.  With floor, values which are closer to
the next lower bound do not appear in the next bin
where they would more naturally go.

Modification:

Change Math.floor to Math.round.

Result:

The desired grouping effect.

Target: master
Acked-by: Paul