authz: return a list of allowed but unasserted roles
Motivation:

To provide a good user experience, it would be useful for dCache-View to
somehow show which roles the user could potentially assert but currently
does not.  For example, this would allow the GUI to give admin users the
possibility to enable their admin role but without showing this option
to any non-admin users.

Currently it is not possible for dCacheView to determine which roles
the user could potentially assert.

Modification:

Return a list of unasserted roles: these are roles that the user would
be authorized to assert but the user did not request.

Update frontend to list those roles that are not asserted.  This allows
a client to make a REST call to discover which roles a user could
assert, in addition to those roles that have been asserted.

Result:

Clients can now discover which roles the user may assert.

Target: master
Require-notes: yes
Require-book: yes
Patch: https://rb.dcache.org/r/10291/
Acked-by: Tigran Mkrtchyan