dcache-restful-api:  add RESTful billing service and support
This patch combines RB 10149, 10151, 10152 and 10153.

Motivation:

Add RESTful frontend billing info service which does the following:

    Delivers the billing records for a given pnfsid to an authorized client.
    Delivers histogram data corresponding to the monitoring plots currently
		exposed via webadmin.

The first is entirely new functionality; the second already exists, but is
keyed to supporting JAIDA-style histograms.

Modifications:

Backend:

(a) Addition of a special cell message type, and a special receiver,
       for billing record requests.
(b) Addition of two views to the billing data to support the querying
       of records for a given pnfsid (liquibase).
(c) Addition of DAO definitions for these views (Datanucleus).
(d) Refactoring of IHistogramData to the db package, in anticipation
       of the eventual deprecation/elimination of the histogram and plot
       packages. This change is responsible for the size of the patch.
(e) Addition of a special cell message type, and a special receiver,
       for time series (histogram) data. This receiver replaces
       the current proxy service interaction, which will be deprecated
       and eventually eliminated.

Patch adds the API resource definition and specialized JSON types,
collector implementation and supporting static utilities for
sending, receiving and processing messages, caching for histogram
data, generation of the necessary messages and their transformation
into the data structures (i.e., BillingData for the records),
and the actual service which uses the collector to retrieve record data
and periodically refresh histogram data.

Target: master
Acked-by: Tigran