common,system-test: fix minor annoyances with roles in system-test
Motivation:

There are two minor configuration issues with system-test as a result of
commit dd5fa1cc.

First, the intended configured behaviour was that users who request a
role they are not authorised to assert should fail the login.
Currently, such roles are silently dropped (which is also valid
behaviour).  This problem stems from gplazma.conf wiring the roles
plugin with 'required' rather than 'requisite'.

Second, currently the "admin" user is not authorized to assert the admin
role.  While not an error per se, this is unfortunate as the webadmin
functional tests require the "admin" user has admin role.

Modification:

Update gPlazma configuration to make "roles" plugin requisite.

Remove gid 1000 as representing admin role and use the default value
(gid 0) to represent users who may assert the admin role.  To achieve
this, the "normal" user should be created as a member of two group.
This, in turn, required extending the utility for writing and updating
kpwd files to support membership of multiple groups.

Result:

In system-test, attempting to assert an unauthorised role fails the
login.

The "admin" user (in addition to the dCache user with the same name as
that of the developer's account) can assert the "admin" role.

The kpwd utilities (dcache kpwd dcuseradd|dcusermode) have been updated
to accept a comma-separated list of gids.

Target: master
Require-notes: no
Require-book: no
Patch: https://rb.dcache.org/r/10304/
Acked-by: Tigran Mkrtchyan